% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisar.R
\name{elisa.analyse}
\alias{elisa.analyse}
\alias{elisa.analyze}
\title{Draws the standard curve of the elisa analysis}
\usage{
elisa.analyse(df, blank = FALSE, tansform = FALSE, tecan = FALSE,
  ignore.dilution = FALSE, std.key = "STD")

elisa.analyze(df, blank = FALSE, tansform = FALSE, tecan = FALSE,
  ignore.dilution = FALSE, std.key = "STD")
}
\arguments{
\item{df}{dataframe containing the O.D. values and at least the 'id' column}

\item{blank}{whether the blank values (rows with id = 'blank') should be substracted or not from all O.D. values}

\item{tecan}{whether the O.D. values should be log10 transformed or not prior regression}

\item{transform}{whether the O.D. values should be log10 transformed or not prior regression}
}
\value{
A list object containing the standard curve and the modified input dataframe to include the calculated concentrations
}
\description{
Requires ggplot2
}
\details{
A complete example on how to perform an analysis can be found at \url{http://eric.koncina.eu/r/elisar}.
}
\examples{
\dontrun{
library(elisar)
library(ggplot2)

# Import file
e <- elisa.load("od_measure.xls")
e <- elisa.analyze(e)
e <- elisa.analyze(e, blank = TRUE, transform = TRUE)
}

}

