% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisar.R
\name{elisa.load}
\alias{elisa.load}
\title{Import the OD measures from the Tecan sunrise excel sheet}
\usage{
elisa.load(input)
}
\arguments{
\item{input}{path to the input files}
}
\description{
Reads excel files exported from Tecan Sunrise and modified according to the documentation.
}
\details{
Example on how to prepare the Tecan excel sheet to be imported can be found at \url{http://eric.koncina.eu/r/elisar}.
}
\examples{
\dontrun{
library(elisar)

# Import file
df <- elisa.load("od_measure.xls")
}

}

