% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisar.R
\name{elisa.standard}
\alias{elisa.standard}
\title{Draws the standard curve of the elisa analysis}
\usage{
elisa.standard(standard, unit = "pg/ml")
}
\arguments{
\item{standard}{list object (or dataframe) returned by the elisa.analyze function}

\item{concentration}{unit of the standard points (defaults to pg/ml)}
}
\description{
Requires ggplot2
}
\details{
A complete example on how to perform an analysis can be found at \url{http://eric.koncina.eu/r/elisar}.
}
\examples{
\dontrun{
library(elisar)
library(ggplot2)

# Import file
e <- read.tecan("od_measure.xls")
e <- elisa.analyze(e)
elisa.standard(e)
# OR supplying the dataframe
elisa.standard(e$standard)
}

}

