% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisar.R
\name{elisa_analyse}
\alias{elisa_analyse}
\title{Analyse the O.D. values (regression)}
\usage{
elisa_analyse(x, std_key = "^STD", dec = ".", var_in = "value",
  var_out = "estimate", .drop = FALSE)
}
\arguments{
\item{std_key}{a character string specifying the common starting pattern of standard point ids (default = "^STD").}

\item{dec}{a character string used as a decimal separator for the encoded standard concentration values.}

\item{var_in}{a character string used for the OD values (default = "value")}

\item{var_out}{a character string used to name the output columns (default = "estimate")}

\item{.drop}{Should input columns be dropped? (default = `FALSE`)}

\item{.data}{dataframe containing at least the value and id columns (with O.D. values and sample identifiers).}

\item{value}{a character string specifying the column containing the O.D. values (default = "value").}
}
\value{
A dataframe including the calculated concentrations, standard deviation and wether the value is in the range of the standard curve.
}
\description{
Performs a 4-PL regression of the standard values and converts the O.D. into concentration values.
}
\details{
A complete example on how to perform an analysis can be found at \url{https://github.com/koncina/elisar}.
}
\examples{
\dontrun{
library(elisar)
# Import file
e <- read_plate("od_measure.xls")
elisa_analyse(e)
elisa_analyse(e, .drop = TRUE)
}

}
