% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{elisar_example}
\alias{elisar_example}
\title{Get path to elisar example}
\usage{
elisar_example()
}
\description{
elisar provides an example file in its `inst/extdata`
directory. This function returns the path to the file.
}
\examples{
elisar_example()
}
