% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_elements.R
\name{extract_elements}
\alias{extract_elements}
\title{Extract elements from excel files}
\usage{
extract_elements(path, sheet = 1, na = "")
}
\arguments{
\item{path}{vector containing the path(s) to the input file(s)}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an integer (the position of the sheet) (See `readxl::read_excel` help page).}

\item{na}{Character vector of strings to use for missing values. (See `readxl::read_excel` help page)}
}
\description{
Extract OD data, layout and ID tables from an excel file. Returns a list containing the plates as a matrix and the ID tables as `data.frame`.
}
\details{
Example on how to prepare the excel file can be found at \url{https://github.com/koncina/elisar}.
}
\examples{
\dontrun{
library(elisar)

# Import file(s)
example_file <- system.file("extdata", "example.xls", package="elisar")
extract_elements(example_file, 1)
extract_elements(example_file, 2)
}

}
