% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisar.R
\name{fit_4pl}
\alias{fit_4pl}
\title{Get the 4PL model generated by drc::drm}
\usage{
fit_4pl(.data, concentration = "concentration", od = "od")
}
\arguments{
\item{.data}{data frame (generated by `extract_standard()`) containing the theoretical concentrations and measured O.D. values.}

\item{concentration}{concentration column}

\item{od}{O.D. column}
}
\value{
An object of class `drc`. Have a look at `?drc::drm` for more informations.
}
\description{
Send the standard curve to drc::drm to generate the 4 parameter logistic fit
}
