% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisar.R
\name{get_standard}
\alias{get_standard}
\title{Extract the standard points}
\usage{
get_standard(.df, .key = "STD", .od = value, .dose = .dose,
  .drop = TRUE)
}
\arguments{
\item{.df}{dataframe containing at least the od and id columns (with O.D. values and sample identifiers).}

\item{.key}{a character string specifying the common starting pattern of standard point ids (default = "STD").}

\item{.od}{the column containing the od values (default = value).}

\item{.dose}{the column that will contain the extracted concentration values}

\item{.drop}{A boolean specifying whether additional columns should be dropped (default) or not.}
}
\value{
A dataframe containing the standard with the extracted concentration value (conc).
}
\description{
Filter out the standard point values according the standard point identification pattern (default = "STD").
}
