% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{list_elements}
\alias{list_elements}
\title{List plate elements in excel files}
\usage{
list_elements(path, na = "")
}
\arguments{
\item{path}{vector containing the path to the excel files.}
}
\value{
A tibble listing the found elements: element (plate or id), position (row, col), format (6, 12, 24, 48, 96 well plate), range (cellranger::cell_limits S3 object to be used in read_excel)
}
\description{
List all plate (data and layout) and associated extended id tables.
}
