% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisar.R
\name{od_to_concentration}
\alias{od_to_concentration}
\title{Analyse the O.D. values (regression)}
\usage{
od_to_concentration(id, value, std_key = "^STD", dec = ".")
}
\arguments{
\item{id}{a character vector containing the identifiers.}

\item{value}{a numerical vector containing the O.D. values.}

\item{std_key}{a character string specifying the common starting pattern of standard point ids (default = "^STD").}

\item{dec}{a character string used as a decimal separator for the encoded standard concentration values.}

\item{.drop}{Should input columns be dropped? (default = `TRUE`)}
}
\value{
A numerical vector containing the calculated concentrations.
}
\description{
Performs a 4-PL regression of the standard values and converts the O.D. into concentration values.
}
\details{
A complete example on how to perform an analysis can be found at \url{https://github.com/koncina/elisar}.
}
\examples{
\dontrun{
library(tidyverse)
library(elisar)

read_plate("od_measure.xls") \%>\%
  mutate(concentration = get_concentration(id, value))
}

}
