% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{read.plate}
\alias{read.plate}
\title{Import the OD/luminescence/fluorescence measures from plate reader exported excel sheets (like Fluostar Optima, or Tecan Sunrise)}
\usage{
read.plate(input, layout = NA, checksum = "md5")
}
\arguments{
\item{input}{vector containing the path(s) to the input file(s)}

\item{layout}{If NA, layout sheets are auto-detected in each input file. Setting layout with paths to layout files overrides the autodetection. The vector must be of same size as input or equal to 1 (overrides all input elements with the same layout).}

\item{checksum}{a character string specifying the algorithm to calculate the data checksum (defaults to "md5"). checksum can be one of c("md5", "sha1", "crc32", "sha256", "sha512").
The checksum is stored in the attributes.}
}
\description{
Reads excel files exported from the MARS software (\url{http://www.bmglabtech.com/en/products/software/mars-data-analysis/}) or Tecan Magellan (\url{http://lifesciences.tecan.com/products/software/magellan_data_analysis_software}) and modified to include the plate layout (see documentation).
}
\details{
Example on how to prepare the excel file can be found at \url{http://eric.koncina.eu/r/elisar}.
}
\examples{
\dontrun{
library(elisar)

# Import file(s)
df <- read.plate("od_measure.xls")
df <- read.plate(c("od_measure1.xls", "od_measure2.xls"))
df <- read.plate(c("od_measure1.xls", "od_measure2.xls"), layout = c(NA, "layout2.xls"))
}

}
