% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{read_elements}
\alias{read_elements}
\title{Read plate elements in excel files}
\usage{
read_elements(.df, na = "")
}
\arguments{
\item{.df}{tibble generated by list_elements().}
}
\value{
A nested tibble containing the data elements: is_layout (boolean), format (6, 12, 24, 48, 96 well plate), data (nested tibble containing the tidy data)
}
\description{
Read all plate (data and layout) and associated extended id tables.
}
