% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_df.R
\name{read_plate}
\alias{read_plate}
\title{Import the OD/luminescence/fluorescence measures from plate reader exported excel sheets (like Fluostar Optima, or Tecan Sunrise)}
\usage{
read_plate(path, na = "")
}
\arguments{
\item{path}{vector containing the path(s) to the input file(s)}

\item{na}{character vector of strings to use for missing values (see readxl::read_excel)}
}
\description{
Reads excel files exported from the MARS software (\url{http://www.bmglabtech.com/en/products/software/mars-data-analysis/}) or Tecan Magellan (\url{http://lifesciences.tecan.com/products/software/magellan_data_analysis_software}) and modified to include the plate layout (see documentation).
}
\details{
Example on how to prepare the excel file can be found at \url{https://github.com/koncina/elisar}.
}
\examples{
\dontrun{
library(elisar)

# Import file(s)
df <- read_plate("od_measure.xls")
df <- read_plate(c("od_measure1.xls", "od_measure2.xls"))
}

}
