% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisar.R
\name{elisa.analyse}
\alias{elisa.analyse}
\alias{elisa.analyze}
\title{Draws the standard curve of the elisa analysis}
\usage{
elisa.analyse(.df, ..., multi.regression = TRUE)

elisa.analyze(.df, ..., multi.regression = TRUE)
}
\arguments{
\item{.df}{dataframe containing the O.D. values and at least the 'id' column}

\item{multi.regression}{perform a regression on each individual file when a multiple files dataset is supplied}

\item{blank}{substract blank values (id = 'blank') from all O.D. values}

\item{log10}{transform O.D. values before regression}

\item{search}{for standard point values by ids starting with std.key (default = "STD")}

\item{ignore.dilution}{ignore the dilution column even if present for the concentration calculation}

\item{tecan}{fix bad Tecan O.D. values (>1000)}
}
\value{
A list object containing the standard curve and the modified input dataframe to include the calculated concentrations
}
\description{
Requires ggplot2
}
\details{
A complete example on how to perform an analysis can be found at \url{http://eric.koncina.eu/r/elisar}.
}
\examples{
\dontrun{
library(elisar)
library(ggplot2)

# Import file
e <- elisa.load("od_measure.xls")
e <- elisa.analyze(e)
e <- elisa.analyze(e, blank = TRUE, transform = TRUE)
}

}

