% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisar.R
\name{elisa.analyse}
\alias{elisa.analyse}
\alias{elisa.analyze}
\title{Analyse the O.D. values (regression)}
\usage{
elisa.analyse(.df, ..., multi.regression = TRUE)

elisa.analyze(.df, ..., multi.regression = TRUE)
}
\arguments{
\item{.df}{dataframe containing at least the od and id columns (with O.D. values and sample identifiers).}

\item{multi.regression}{a logical value indicating whether the data set should be split by filename before the regression (when multiple files are loaded with 'elisa.load').}

\item{blank}{a logical value indicating whether blank values (id = 'blank') should be substracted from all O.D. values.}

\item{transform}{a logical value indicating whether O.D. values should be log10 transformed before the regression.}

\item{std.key}{a character string specifying the common starting pattern of standard point ids (default = "STD").}

\item{dilution.column}{NULL or character string specifying a column to be used for the sample dilution factors (default = NULL).}

\item{tecan}{a logical value indicating whether bad Tecan O.D. values (>1000) should be fixed.}
}
\value{
A list object containing the standard curve and the modified input dataframe to include the calculated concentrations.
}
\description{
Performs a 4-PL regression of the standard values and converts the O.D. into concentration values.
}
\details{
A complete example on how to perform an analysis can be found at \url{http://eric.koncina.eu/r/elisar}.
}
\examples{
\dontrun{
library(elisar)
library(ggplot2)

# Import file
e <- elisa.load("od_measure.xls")
e <- elisa.analyze(e)
e <- elisa.analyze(e, blank = TRUE, transform = TRUE)
}

}

