% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisar.R
\name{elisa.load}
\alias{elisa.load}
\title{Import the OD measures from the Tecan sunrise excel sheet}
\usage{
elisa.load(input, checksum = NULL)
}
\arguments{
\item{input}{vector containing the path(s) to the input file(s)}

\item{checksum}{a character string specifying the algorithm to calculate the data checksum (defaults to NULL). checksum can be one of c("md5", "sha1", "crc32", "sha256", "sha512").}
}
\description{
Reads excel files exported from Tecan Sunrise and modified according to the documentation.
}
\details{
Example on how to prepare the Tecan excel sheet to be imported can be found at \url{http://eric.koncina.eu/r/elisar}.
}
\examples{
\dontrun{
library(elisar)

# Import file
df <- elisa.load("od_measure.xls")
df <- elisa.load(c("od_measure1.xls", "od_measure2.xls"))
}

}

