% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisar.R
\name{extract_standard}
\alias{extract_standard}
\title{Extract the standard curve}
\usage{
extract_standard(.data, x = "x", y = "y", std_key = "^STD", dec = ".")
}
\arguments{
\item{.data}{dataframe containing at least the value and id columns (with O.D. values and sample identifiers).}

\item{x}{name of the concentration column}

\item{y}{name of the O.D. column}

\item{std_key}{a character string specifying the common starting pattern of standard point ids (default = "^STD").}

\item{dec}{a character string used as a decimal separator for the encoded standard concentration values.}
}
\value{
A data frame containing the standard curve
}
\description{
Parse the id for the standard curve key and extract the concentration and associated O.D. values.
}
